acceptance: extract C test into file
Improve maintainability by putting the C source code used in TestDockerC
into its own .c file.

Release note: None
