+++ pkg/acceptance/testdata/c/Makefile
@@ -1,0 +1,3 @@
+CC += -std=c99
+CPPFLAGS += -I/usr/include/postgresql
+LDLIBS += -lpq -lpqtypes
